+++ frontend/src/components/ui/form.tsx
@@ -24,7 +24,7 @@ type FormFieldContextValue<
 }
 
 const FormFieldContext = React.createContext<FormFieldContextValue>(
-  {} as FormFieldContextValue
+  {} as FormFieldContextValue,
 )
 
 const FormField = <
@@ -68,7 +68,7 @@ type FormItemContextValue = {
 }
 
 const FormItemContext = React.createContext<FormItemContextValue>(
-  {} as FormItemContextValue
+  {} as FormItemContextValue,
 )
 
 function FormItem({ className, ...props }: React.ComponentProps<"div">) {
@@ -153,8 +153,9 @@ function FormMessage({ className, ...props }: React.ComponentProps<"p">) {
   )
 }
 
+// eslint-disable-next-line react-refresh/only-export-components
+export { useFormField }
 export {
-  useFormField,
   Form,
   FormItem,
   FormLabel,