+++ frontend/vite.config.ts
@@ -1,23 +1,21 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react-swc'
-import { viteEnvs } from 'vite-envs'
-import tailwindcss from '@tailwindcss/vite'
+import { defineConfig } from "vite"
+import react from "@vitejs/plugin-react-swc"
+import { viteEnvs } from "vite-envs"
+import tailwindcss from "@tailwindcss/vite"
 import path from "path"
-
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [
     viteEnvs({
-      declarationFile: '.env.example',
+      declarationFile: ".env.example",
     }),
-   react(),
+    react(),
     tailwindcss(),
   ],
-    resolve: {
+  resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
     },
   },
-
 })