+++ terraform/shitty-aws-architecture/06-s3.tf
@@ -81,3 +81,23 @@ resource "aws_s3_bucket_cors_configuration" "images" {
     max_age_seconds = 3000
   }
 }
+
+# S3 Bucket Policy for Public Read Access
+resource "aws_s3_bucket_policy" "images_public_read" {
+  bucket = aws_s3_bucket.images.id
+
+  policy = jsonencode({
+    Version = "2012-10-17"
+    Statement = [
+      {
+        Sid       = "PublicReadGetObject"
+        Effect    = "Allow"
+        Principal = "*"
+        Action    = "s3:GetObject"
+        Resource  = "${aws_s3_bucket.images.arn}/*"
+      }
+    ]
+  })
+
+  depends_on = [aws_s3_bucket_public_access_block.images]
+}