+++ terraform/shitty-aws-architecture/08-ec2_ami_builder.tf
@@ -8,7 +8,7 @@ resource "aws_instance" "ami_builder" {
   vpc_security_group_ids = [aws_security_group.app_server.id]
   subnet_id              = aws_subnet.public[0].id # Needs public internet to pull docker images
 
-  user_data = base64encode(templatefile("${path.module}/user_data_ami_builder.sh", {
+  user_data_base64 = base64encode(templatefile("${path.module}/user_data_ami_builder.sh", {
     AWS_REGION              = var.aws_region
     DB_HOST                 = aws_db_instance.main.address
     DB_PORT                 = aws_db_instance.main.port
@@ -24,6 +24,7 @@ resource "aws_instance" "ami_builder" {
     DOCKER_IMAGE_FRONTEND   = "ghcr.io/taldoflemis/sora-henkan/frontend:latest"
     API_DOMAIN              = var.api_domain
     ALB_DNS_NAME            = aws_lb.app.dns_name
+    AWS_BUCKET_ENDPOINT     = "https://${aws_s3_bucket.images.bucket}.s3.${var.aws_region}.amazonaws.com/"
   }))
 
   tags = merge(